deployments/requirements.txt
torch
click
ovmsclient
numpy
transformers
requests
scipy
optimum-intel
openvino
onnx
onnxruntime
tritonclient[all]
